Added dialogs and fix styling

+++ src/app/pages/Task/task-list/task-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, inject, Input, OnInit } from '@angular/core';
 import { TaskService } from '../../../services/Task/task.service';
 import { NotificationService } from '../../../services/notification/notification.service';
 import { MatIconModule } from '@angular/material/icon';
@@ -7,6 +7,10 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatBadgeModule } from '@angular/material/badge';
 import { MatCardModule } from '@angular/material/card';
 import { Router } from '@angular/router';
+import { LoaderService } from '../../../services/loader/loader.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../dialogs/confirm-dialog/confirm-dialog.component';
+import { EditTaskDialogComponent } from '../dialogs/EditTaskDialog/EditTaskDialog.component';
 export interface Task {
   _id: string;
   title: string;
@@ -30,26 +34,27 @@ export interface Task {
 })
 export class TaskListComponent implements OnInit {
   tasks: Task[] = [];
+  readonly dialog = inject(MatDialog);
 
   constructor(
     private taskService: TaskService,
-    private router: Router,
-    private notificationService: NotificationService
+    private notificationService: NotificationService,
+    private loaderService: LoaderService
   ) {}
 
   ngOnInit(): void {
     this.fetchTasks();
 
- 
     this.taskService.getTasksUpdatedListener().subscribe(() => {
-      this.fetchTasks(); 
+      this.fetchTasks();
     });
   }
 
-
   fetchTasks(): void {
+    this.loaderService.showLoader();
     this.taskService.getAllTasks().subscribe(
       (response: Task[]) => {
+        this.loaderService.hideLoader();
         this.tasks = response.reverse();
       },
       (error) => {
@@ -57,6 +62,7 @@ export class TaskListComponent implements OnInit {
           'Failed to load tasks',
           'error'
         );
+        this.loaderService.hideLoader();
       }
     );
   }
@@ -65,9 +71,9 @@ export class TaskListComponent implements OnInit {
     return task._id;
   }
 
-
   // Toggle task completion
   toggleTaskCompletion(taskId: string, currentStatus: boolean): void {
+    this.loaderService.showLoader();
     if (taskId) {
       const updatedStatus = !currentStatus;
 
@@ -75,12 +81,14 @@ export class TaskListComponent implements OnInit {
         .toggleTaskStatus(taskId, { completed: updatedStatus })
         .subscribe(
           (response) => {
+            this.loaderService.hideLoader();
             this.notificationService.showNotification(
               'Task status updated',
               'success'
             );
           },
           (error) => {
+            this.loaderService.hideLoader();
             this.notificationService.showNotification(
               'Error updating task status',
               'error'
@@ -88,6 +96,7 @@ export class TaskListComponent implements OnInit {
           }
         );
     } else {
+      this.loaderService.hideLoader();
       this.notificationService.showNotification(
         'Task ID is not defined',
         'error'
@@ -96,26 +105,51 @@ export class TaskListComponent implements OnInit {
   }
 
   // Edit task
-  editTask(taskId: string): void {
-    this.router.navigate(['/tasks/', taskId]);
+  openEditTaskDialog(taskId: string): void {
+    this.loaderService.showLoader();
+    const dialogRef = this.dialog.open(EditTaskDialogComponent, {
+      width: '800px',
+      panelClass: 'app-EditTaskDialog',
+      data: { taskId },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.loaderService.hideLoader();
+        console.log('Dialog closed with result:', result);
+      }
+    });
   }
+
 
   // Delete task
   deleteTask(taskId: string): void {
-    this.taskService.deleteTask(taskId).subscribe(
-      () => {
-        this.tasks = this.tasks.filter((task) => task._id !== taskId);
-        this.notificationService.showNotification(
-          'The task has been deleted',
-          'success'
-        );
-      },
-      (error) => {
-        this.notificationService.showNotification(
-          'Failed to delete task',
-          'error'
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      panelClass: 'app-confirm-dialog',
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.loaderService.showLoader();
+        this.taskService.deleteTask(taskId).subscribe(
+          () => {
+            this.tasks = this.tasks.filter((task) => task._id !== taskId);
+            this.loaderService.hideLoader();
+            this.notificationService.showNotification(
+              'The task has been deleted',
+              'success'
+            );
+          },
+          (error) => {
+            this.loaderService.hideLoader();
+            this.notificationService.showNotification(
+              'Failed to delete task',
+              'error'
+            );
+          }
         );
       }
-    );
+    });
   }
 }
